+++ CHANGELOG.md
@@ -85,6 +85,8 @@
 * Make the Move actions available regardless of how user navigated to the record. Refs UIIN-2518.
 * Sorting on the Instance's Holdings item table is not working, part 2. Fixes UIIN-2565.
 * Promote a local instance to become a shared instance. Refs UIIN-2460.
+* Consortial Central Tenant: Suppress Holdings detail view Action menu. Refs UIIN-2524.
+* Consortial Central Tenant: Suppress Item detail view Action menu. Refs UIIN-2525.
 * Consortial Central Tenant: Hide Instance Action Menu New Request option. Refs UIIN-2572.
 
 ## [9.4.11](https://github.com/folio-org/ui-inventory/tree/v9.4.11) (2023-08-02)